backend/requirements.txt
fastapi
uvicorn
sqlalchemy[asyncio]
asyncpg
aioredis
redis
celery
pydantic
pydantic-settings
python-dotenv
openai
httpx
python-multipart
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
psycopg2-binary
alembic
pytest
pytest-asyncio
